one attribute per file
