+++ src/Neo.SmartContract.Framework/Attributes/AuthorAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Neo.SmartContract.Framework.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class AuthorAttribute : ManifestExtraAttribute
+    {
+        public AuthorAttribute(string value) : base(AttributeType[nameof(AuthorAttribute)], value)
+        {
+        }
+    }
+}
