+++ src/Neo.SmartContract.Framework/Attributes/DescriptionAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Neo.SmartContract.Framework.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class DescriptionAttribute : ManifestExtraAttribute
+    {
+        public DescriptionAttribute(string value) : base(AttributeType[nameof(DescriptionAttribute)], value)
+        {
+        }
+    }
+}
