+++ src/Neo.SmartContract.Framework/Attributes/EmailAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Neo.SmartContract.Framework.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class EmailAttribute : ManifestExtraAttribute
+    {
+        public EmailAttribute(string value) : base(AttributeType[nameof(EmailAttribute)], value)
+        {
+        }
+    }
+}
